Split out classes, simplify logic, minor design changes
